+++ Source/Fluxor.Blazor.Web.ReduxDevTools/ReduxDevToolsMiddlewareOptions.cs
@@ -41,12 +41,26 @@ namespace Fluxor.Blazor.Web.ReduxDevTools
 		public ReduxDevToolsMiddlewareOptions UseNewtonsoftJson(
 			Func<IServiceProvider, JsonSerializerSettings> getSettings = null)
 		{
-			FluxorOptions.Services.AddScoped<IJsonSerialization, NewtonsoftSerialization>(
-				sp =>
+			FluxorOptions.Services.AddScoped<IJsonSerialization, NewtonsoftJsonAdapter>(
+				(Func<IServiceProvider, NewtonsoftJsonAdapter>)(				sp =>
 				{
 					JsonSerializerSettings settings = getSettings?.Invoke(sp);
+
+/* Unmerged change from project 'Fluxor.Blazor.Web.ReduxDevTools (netstandard2.1)'
+Before:
 					return new NewtonsoftSerialization(settings);
-				});
+After:
+					return new Serialization.NewtonsoftSerialization(settings);
+*/
+
+/* Unmerged change from project 'Fluxor.Blazor.Web.ReduxDevTools (net5.0)'
+Before:
+					return new NewtonsoftSerialization(settings);
+After:
+					return new Serialization.NewtonsoftSerialization(settings);
+*/
+					return (NewtonsoftJsonAdapter)new NewtonsoftJsonAdapter(settings);
+				}));
 			return this;
 		}
 
@@ -58,12 +72,26 @@ namespace Fluxor.Blazor.Web.ReduxDevTools
 		public ReduxDevToolsMiddlewareOptions UseSystemTextJson(
 			Func<IServiceProvider, JsonSerializerOptions> getOptions = null)
 		{
-			FluxorOptions.Services.AddScoped<IJsonSerialization, SystemTextSerialization>(
-				sp =>
+			FluxorOptions.Services.AddScoped<IJsonSerialization, SystemTextJsonAdapter>(
+				(Func<IServiceProvider, SystemTextJsonAdapter>)(				sp =>
 				{
 					JsonSerializerOptions jsonOptions = getOptions?.Invoke(sp);
+
+/* Unmerged change from project 'Fluxor.Blazor.Web.ReduxDevTools (netstandard2.1)'
+Before:
 					return new SystemTextSerialization(jsonOptions);
-				});
+After:
+					return new Serialization.SystemTextSerialization(jsonOptions);
+*/
+
+/* Unmerged change from project 'Fluxor.Blazor.Web.ReduxDevTools (net5.0)'
+Before:
+					return new SystemTextSerialization(jsonOptions);
+After:
+					return new Serialization.SystemTextSerialization(jsonOptions);
+*/
+					return (SystemTextJsonAdapter)new SystemTextJsonAdapter(jsonOptions);
+				}));
 			return this;
 		}
 	}